pyproject.toml
[tool.poetry]
name = "interviewai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
pypdf = "^4.1.0"
tiktoken = "^0.6.0"
flask-socketio = "^5.3.6"
flask-cors = "^4.0.0"
pyjwt = "^2.8.0"
pydantic = "1.10.8"
deepgram-sdk = "3.*"
#google-cloud-logging = "^3.10.0"
#google-cloud-profiler = "^4.1.0"
google-cloud-firestore = "^2.16.0"
google-cloud-storage = "^2.16.0"
google-auth = "^2.29.0"
functions-framework = "3.5.0"

firebase-admin = "^6.5.0"
stripe = "^8.10.0"
svix = "^1.21.0"
boto3 = "^1.34.79"
azure-cognitiveservices-vision-computervision = "^0.9.0"
pymupdf = "^1.24.1"
firebase = "^4.0.1"
flask = "^2.3.3"
python-docx = "^1.1.0"
pinecone-client = "^3.2.2"
langchain = "^0.1.16"
openai = "^1.17.1"
langchain-openai = "^0.1.3"
pillow = "^10.3.0"
opentelemetry-api = "^1.24.0"
opentelemetry-sdk = "^1.24.0"
opentelemetry-exporter-gcp-trace = "^1.6.0"
gunicorn = "^22.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
